\name{bathy.colors}
\alias{bathy.colors}
\title{Bathymetric data color scheme}
\description{
  Generates a color ramp from light grey to blue.
}
\usage{
  bathy.colors(n)
}
\arguments{
  \item{n}{The number of colors to generate}
}
\value{
  A vector of \code{n} colors linearly spaced from light grey to pure blue.
}
\author{Benjamin Jones \email{btjones16@gmail.com}}
  