% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_builder_functions.R
\name{buildTree}
\alias{buildTree}
\title{buildtree}
\usage{
buildTree(root, graph, remove_reverse = F)
}
\arguments{
\item{root}{ipsum...}

\item{graph}{ipsum...}
}
\value{
ipsum...
}
\description{
buildtree
}
