% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_credentials.r
\name{check_credentials}
\alias{check_credentials}
\title{Ésta es una función para chequear si las credenciales de observe están en el sistema. Si no, las pide y da la opción de escribirlas.
Las escribirá en la ruta del usuario}
\usage{
check_credentials()
}
\description{
Ésta es una función para chequear si las credenciales de observe están en el sistema. Si no, las pide y da la opción de escribirlas.
Las escribirá en la ruta del usuario
}
\examples{
check_credentials()


}
