% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{checkInputs}
\alias{checkInputs}
\title{checkInputs}
\usage{
checkInputs(measurments, targets)
}
\arguments{
\item{measurments}{ipsum...}

\item{targets}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
