% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_pathway.R
\name{compress_transporters}
\alias{compress_transporters}
\title{\code{compress_transporters}}
\usage{
compress_transporters(sub_network_nocofact)
}
\arguments{
\item{sub_network_nocofact}{ipsum...}
}
\value{
ipsum...
}
\description{
This function removes cofactors from reaction networks generate by the model_to_pathway_sif function
}
