% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metactivity_core_functions.R
\name{condense_metabolite_set}
\alias{condense_metabolite_set}
\title{condense_metabolite_set}
\usage{
condense_metabolite_set(
  reaction_set_list,
  condense_sign = T,
  condense_compartments = T,
  ignore_sign = F
)
}
\arguments{
\item{reaction_set_list}{ipsum...}

\item{condense_sign}{ipsum...}

\item{condense_compartments}{ipsum...}

\item{ignore_sign}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
