% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.environment.r
\name{create.environment}
\alias{create.environment}
\title{Ésta es una función para instalar el modulo de copernicus marine}
\usage{
create.environment()
}
\description{
Ésta es una función para instalar el modulo de copernicus marine
}
\examples{
create.environment()


}
