% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_copernicus.r
\name{download_copernicus}
\alias{download_copernicus}
\title{This is a function to download data from Copernicus.}
\usage{
download_copernicus(
  yearstart,
  monthstart,
  yearend,
  monthend,
  minlon,
  maxlon,
  minlat,
  maxlat,
  mindepth = 0,
  maxdepth = 4,
  product = NULL,
  outputdir = NULL
)
}
\arguments{
\item{yearstart}{First year of time series}

\item{monthstart}{First month of time series}

\item{yearend}{Last year of time series}

\item{monthend}{Last month on fime series}

\item{minlon}{minimum longitude}

\item{maxlon}{maximum longitude}

\item{minlat}{minimum latitude}

\item{maxlat}{maximum latitude}

\item{mindepth}{mimimum depth}

\item{maxdepth}{maximum depth}

\item{product}{product name (from data(productlist))}

\item{outputdir}{path to directore to save data.}

\item{username}{copernicus username}

\item{password}{copernicus password}
}
\description{
This is a function to download data from Copernicus.
}
\examples{
dir.create('./test')
yearstart=2010
monthstart=1
yearend=2010
monthend=1
minlon=-25
maxlon=-20
minlat=35
maxlat=40
mindepth=0
maxdepth=4
product=c('mchl')
outputdir='./test'
productlist<-read.csv('P:/Pacifico/Rwork/Rprojects/ocean/mercatorproducts.csv')
download_copernicus(yearstart,monthstart,
                    yearend,monthend,
                   minlon,maxlon,minlat,maxlat,mindepth,maxdepth, 
                    product,outputdir,username,
                    password)
}
\keyword{download,copernicus}
