% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressed_genes.R
\docType{data}
\name{expressed_genes}
\alias{expressed_genes}
\title{expressed_genes}
\format{
An object of class \code{data.frame} with 28394 rows and 2 columns.
}
\usage{
data(expressed_genes)
}
\description{
expressed_genes
}
\examples{
data(expressed_genes)

}
\references{
{
  https://pubmed.ncbi.nlm.nih.gov/36539415/
}
}
\keyword{expressed_genes}
