% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_builder_functions.R
\name{forestMaker}
\alias{forestMaker}
\title{forestMaker}
\usage{
forestMaker(
  molecule_names,
  reaction_network,
  branch_length = c(1, 1),
  remove_reverse = F
)
}
\arguments{
\item{molecule_names}{ipsum...}

\item{reaction_network}{ipsum...}
}
\value{
ipsum...
}
\description{
forestMaker
}
