% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reaction_network.R
\name{generate_sifAtt}
\alias{generate_sifAtt}
\title{generate_sifAtt}
\usage{
generate_sifAtt(
  network_and_attributes,
  t_table,
  scores_df,
  column_index,
  rbrewer_plalette_name = "RdBu"
)
}
\arguments{
\item{network_and_attributes}{lorem ipsum...}

\item{t_table}{lorem ipsum...}

\item{scores_df}{lorem ipsum...}

\item{column_index}{lorem ipsum...}

\item{rbrewer_plalette_name}{lorem ipsum...}
}
\value{
lorem ipsum...
}
\description{
generate_sifAtt
}
