\docType{methods}
\name{get.depth}
\alias{get.depth}
\alias{get.depth,fvcom.grid-method}
\title{Get the depth at each node in the grid.}
\arguments{
  \item{grid}{A \code{fvcom.grid} instance.}
}
\value{
  A vector of length \code{get.nnodes(grid)} with the depth at
  each node.
}
\description{
  Get the depth at each node in the grid.
}

