\docType{methods}
\name{get.elems}
\alias{get.elems}
\alias{get.elems,fvcom.grid-method}
\title{Get the indices of the element vertices in the grid.}
\arguments{
  \item{grid}{A \code{fvcom.grid} instance.}
}
\value{
  A \code{data.frame} with \code{v1}, \code{v2}, and
  \code{v3} elements that correspond to the indices of the
  nodes at vertices 1, 2, and 3.
}
\description{
  Get the indices of the element vertices in the grid.
}

