\docType{methods}
\name{get.nelems}
\alias{get.nelems}
\alias{get.nelems,fvcom.grid-method}
\title{Get the number of elements in the grid.}
\arguments{
  \item{grid}{A \code{fvcom.grid} instance.}
}
\value{
  The number of elements in \code{grid}.
}
\description{
  Get the number of elements in the grid.
}

