\docType{methods}
\name{get.nnodes}
\alias{get.nnodes}
\alias{get.nnodes,fvcom.grid-method}
\title{Get the number of nodes in the grid.}
\arguments{
  \item{grid}{A \code{fvcom.grid} instance.}
}
\value{
  The number of nodes in \code{grid}.
}
\description{
  Get the number of nodes in the grid.
}

