\docType{methods}
\name{get.nodes}
\alias{get.nodes}
\alias{get.nodes,fvcom.grid-method}
\title{Get the values of the nodes in the grid.}
\arguments{
  \item{grid}{A \code{fvcom.grid} instance.}
}
\value{
  A \code{data.frame} with \code{x}, \code{y}, and \code{h}
  elements that correspond to the x, y, and depth of each
  node.
}
\description{
  Get the values of the nodes in the grid.
}

