\docType{methods}
\name{get.proj}
\alias{get.proj}
\alias{get.proj,fvcom.grid-method}
\title{Get the value of the projection string.}
\arguments{
  \item{grid}{A \code{fvcom.grid} instance.}
}
\value{
  The value of the projection string specified when the
  grid was created.
}
\description{
  Get the value of the projection string.
}

