% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_copernicus.r
\name{get_copernicus}
\alias{get_copernicus}
\title{This is a function to interpolate copernicus data, given the date, longitude and latitude.}
\usage{
get_copernicus(yy, mm, dd, lon, lat, product, folder, level = 1)
}
\arguments{
\item{yy}{vector with the year component of the date}

\item{mm}{vector with the month component of the date}

\item{dd}{vector with the day component of the date}

\item{lon}{vector of longitudes}

\item{lat}{vector of latitudes}

\item{folder}{path to folder with copernicus data}

\item{level}{depth level}
}
\description{
This is a function to interpolate copernicus data, given the date, longitude and latitude.
}
\examples{
See temperature at around 15m deep in several locations in January
dir.create('./test')
download_copernicus(yearstart=2017,monthstart=1,yearend=2017,
monthend=1,minlon=-15,maxlon=5,minlat=35,maxlat=45,mindepth=10,maxdepth=20,c('mtemp'),'./test',username='username',
password='mypassword')
test<-load_copernicus(2017,1,'mtemp','./test')
test$depth
plot_copernicus(yy=2017,mm=1,dd=12,product='mtemp',folder='./test',level=3,main='Prueba temperatura a 15 m 12/01/2014',xlab='Longitud',ylab='Latitud',log=F,interpolate = T)
testdata<-data.frame(year=rep(2017,5),month=rep(1,5),day=c(1,2,13,20,4),lon=c(-10,-11,-12,-5,3),lat=c(36,38,40,38,38))
testdata$temp15<-get_copernicus(yy=testdata$year,mm=testdata$month,dd=testdata$day,
                                lon=testdata$lon,lat=testdata$lat,product='mtemp',folder='./test/',level=3)
testdata
}
\keyword{get,copernicus}
