% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_metabolite_enrichment.R
\name{get_pure_kegg_ids}
\alias{get_pure_kegg_ids}
\title{\code{get_pure_kegg_ids}}
\usage{
get_pure_kegg_ids(metabolites_col)
}
\arguments{
\item{metabolites_col}{column in a data frame containing metabolite ids}
}
\value{
column in a data frame containing KEGG ids
}
\description{
Function to remove the compartment information and "cpd:" from a column with
metabolites identifiers in order to get pure KEGG ids
}
