% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.copernicuscredentials.r
\name{insert.copernicuscredentials}
\alias{insert.copernicuscredentials}
\title{Ésta es una función para orzar la entrada de nuevas credenciales
incluyendo la ruta a mamba}
\usage{
insert.copernicuscredentials()
}
\description{
Ésta es una función para orzar la entrada de nuevas credenciales
incluyendo la ruta a mamba
}
\examples{
insert.copernicuscredentials()


}
