% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.copernicus.r
\name{install.copernicus}
\alias{install.copernicus}
\title{This is a script to try to make the whole installation...}
\usage{
install.copernicus()
}
\description{
This is a script to try to make the whole installation...
}
\examples{
install.copernicus()
}
\keyword{download,copernicus}
