\name{jet.colors}
\alias{jet.colors}
\title{Jet color ramp}
\description{
  Provides an implementation of a color scheme similar to the Jet
  colormap produced by matplotlib.
}
\usage{
  jet.colors(n)
}
\arguments{
  \item{n}{The number of colors to produce.}
}
\value{
  A vector of colors of length \code{n}.
}
\author{Benjamin Jones \email{btjones16@gmail.com}}
  