% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{limma_res_to_ttop_list}
\alias{limma_res_to_ttop_list}
\title{limma_res_to_ttop_list}
\usage{
limma_res_to_ttop_list(limma_res, comp_names, number, adjust.method = "fdr")
}
\arguments{
\item{limma_res}{output of the runLimma function}

\item{comp_names}{identifiers of the contrasts that were considered in the 
runLimma function}

\item{number}{number of top features to be included in the top table results}

\item{adjust.method}{method to adjust p-values, same as for the topTable limma function}
}
\value{
a list of limma top table results
}
\description{
This function converts the output of the runLimma function into a list of
limma top tables
}
