% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_copernicus.r
\name{load_copernicus}
\alias{load_copernicus}
\title{This is a function to load daily products, downloaded as monthly files from Copernicus by using the function download_copernicus()
The biogeochemical model product before 2012 has several issues and do not work well...(varying long and lat)}
\usage{
load_copernicus(yy, mm, product, folder)
}
\arguments{
\item{yy}{year}

\item{mm}{month}

\item{product}{product name (from data(productlist))}

\item{folder}{folder with copernicus data}
}
\description{
This is a function to load daily products, downloaded as monthly files from Copernicus by using the function download_copernicus()
The biogeochemical model product before 2012 has several issues and do not work well...(varying long and lat)
}
\examples{
dir.create('./test')
download_copernicus(yearstart=2017,monthstart=1,yearend=2017,
monthend=1,minlon=-15,maxlon=5,minlat=35,maxlat=45,mindepth=0,maxdepth=4,c('chl'),'./test',username='username',
password='password')
test<-load_copernicus(2017,1,'chl','./test')
Cholorophylle at the surface (log transformed)
image(test$lon,test$lat,log(test$z[,,1,12]))
}
\keyword{load,copernicus}
