% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{make_df_and_targets_great_again}
\alias{make_df_and_targets_great_again}
\title{\code{make_df_and_targets_great_again}}
\usage{
make_df_and_targets_great_again(df, targets)
}
\arguments{
\item{df}{the measurment n*m dataframe (n is number of omic features, m is number of samples) where columns are ordered by conditions.}

\item{targets}{A n*2 dataframe, where n is the number of samples. First column correspond to samples, second column correspond to conditions.}
}
\value{
A list, first element is the cleaned measurment dataframe and second element is the cleaned target dataframe.
}
\description{
This function Check wether sample names are coherent between the measurment dataframe and the target dataframe. If they are coherent,
the target dataframe rows are reordered to match the column order of the measurment dataframe.
}
