% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reaction_network.R
\name{make_discrete_palette}
\alias{make_discrete_palette}
\title{make_discrete_palette}
\usage{
make_discrete_palette(numeric_vector, rbrewer_plalette_name)
}
\arguments{
\item{numeric_vector}{a metabolic-enzyme set collections, 
as returned by the condense_metabolite_set function}

\item{rbrewer_plalette_name}{numeric, the index of the reaction_set_list_merged corresponding to
the desired penalty}
}
\value{
ipsum lorem...
}
\description{
make_discrete_palette
}
