% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualdownloadcopernicus.r
\name{manualdownloadcopernicus}
\alias{manualdownloadcopernicus}
\title{Function to download data from copernicus, just pasting the text obtained when doing a manual selection and clicking on automate and then copying the commmand line interface tex .}
\usage{
manualdownloadcopernicus(CLItext, outputdir, filename)
}
\arguments{
\item{CLItext}{Text as obtained from marine copernicus query tool}

\item{outputdir}{Dir where file will be stored}

\item{filename}{filename}
}
\description{
Function to download data from copernicus, just pasting the text obtained when doing a manual selection and clicking on automate and then copying the commmand line interface tex .
}
\examples{
textfromweb="subset --dataset-id cmems_mod_glo_phy_my_0.083deg_P1D-m --variable thetao --start-datetime 1993-01-01T00:00:00 --end-datetime 2021-06-30T00:00:00 --minimum-longitude -5.3 --maximum-longitude 2.18 --minimum-latitude 35.07 --maximum-latitude 36.73 --minimum-depth 0.49402499198913574 --maximum-depth 0.49402499198913574"
manualdownloadcopernicus(textfromweb,'test','temp.nc')
}
\keyword{download,copernicus}
