% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_metabolite_enrichment.R
\name{map_pathways_to_metabolites}
\alias{map_pathways_to_metabolites}
\title{\code{map_pathways_to_metabolites}}
\usage{
map_pathways_to_metabolites(metab_df)
}
\arguments{
\item{metab_df}{data frame with metabolites and their target/source enzymes}
}
\value{
data frame containing metabolites and their pathways
}
\description{
Function to map pathways to enzymes and subsequently to metabolites.
}
