% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_table.R
\docType{data}
\name{mapping_table}
\alias{mapping_table}
\title{mapping_table}
\format{
An object of class \code{data.frame} with 2594 rows and 2 columns.
}
\usage{
data(mapping_table)
}
\description{
mapping_table
}
\examples{
data(mapping_table)

}
\references{
{
  https://pubmed.ncbi.nlm.nih.gov/36539415/
}
}
\keyword{mapping_table}
