% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metactivity_core_functions.R
\name{metactivity}
\alias{metactivity}
\title{metactivity}
\usage{
metactivity(
  metabolomic_t_table,
  regulons_df,
  compartment_pattern = "",
  k = 1000
)
}
\arguments{
\item{metabolomic_t_table}{ipsum...}

\item{regulons_df}{ipsum...}

\item{compartment_pattern}{ipsum...}

\item{k}{ipsum...}
}
\value{
ipsum...
}
\description{
Thsi is where the fun happens
}
