% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_pathway.R
\name{model_to_pathway_sif}
\alias{model_to_pathway_sif}
\title{\code{model_to_pathway_sif}}
\usage{
model_to_pathway_sif(
  pathway_to_keep,
  pathways = recon2_redhuman$pathway,
  rxns = recon2_redhuman$rxns,
  rules = recon2_redhuman$rules,
  minmax = recon2_redhuman$minmax,
  stochio = recon2_redhuman$stochio,
  metab = recon2_redhuman$metab,
  mapping = recon2_redhuman$gene_mapping,
  recon1_metabolites = recon2_redhuman$metab_mapping
)
}
\arguments{
\item{pathway_to_keep}{ipsum...}

\item{pathways}{ipsum...}

\item{rxns}{ipsum...}

\item{rules}{ipsum...}

\item{minmax}{ipsum...}

\item{stochio}{ipsum...}

\item{metab}{ipsum...}

\item{mapping}{ipsum...}

\item{recon1_metabolites}{ipsum...}
}
\value{
ipsum...
}
\description{
This function generate a subreaction network in sif format from the recon_redhuman model for a given set of pathways
}
