% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{nicePCA}
\alias{nicePCA}
\title{\code{nicePCA}}
\usage{
nicePCA(
  df,
  targets,
  components = c(1, 2, 3),
  centering = T,
  scaling = F,
  pointSize = 4,
  no_label = FALSE
)
}
\arguments{
\item{df}{the measurment n*m dataframe (n is number of omic features, m is number of samples) where columns are ordered by conditions.}

\item{targets}{A n*2 dataframe, where n is the number of samples. First column correspond to samples, second column correspond to conditions.}

\item{components}{a vector of three integers, corresponding to the components to be plotted}

\item{centering}{a boolean parameter to indicate wether samples should be mean centered}

\item{scaling}{a boolean parameter to indicate wether samples should be scaled (x/variance)}

\item{pointSize}{an integer parameter to indicate the desired point size for the components scatter plots.}
}
\value{
an 3*3 arrangeGrob object containing various graphical representation of the result of a PCA.
}
\description{
This function generate a 3*3 arrangeGrob plot object (that can be subsequently diplayed or saved).
Each cell of the 3*3 plot grid correspond to a specific representation of the result of a principal component analysis performed on a measurment dataframe.
The first input is a n*m data.frame, where n is the number of measured omic features (genes, proteins, metabolites...) and m is the number of samples.
The second input is a basic n*2 target dataframe (such as generated by the generateTarget function), where n is the number of samples.
}
