% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_pathway.R
\name{nitrogen_tracking}
\alias{nitrogen_tracking}
\title{\code{nitrogen_tracking}}
\usage{
nitrogen_tracking(sub_network_nocofact)
}
\arguments{
\item{sub_network_nocofact}{ipsum...}
}
\value{
ipsum...
}
\description{
This function cross link transamination metabolites to track amine groups
}
