\docType{data}
\name{ocean.demo.grid}
\alias{ocean.demo.grid}
\title{An example instance of the \code{fvcom.grid} class.}
\format{A \code{fvcom.grid} instance.}
\description{
  A dataset containing an example of the \code{fvcom.grid}
  class. This example was generated for the purposes of
  demonstrating the plotting functions of this package and
  is not an actual grid. The grid uses bathymetry from
  NOAA's ETOPO1 dataset, which is available at
  (http://www.ngdc.noaa.gov/mgg/global/global.html).
}
\keyword{datasets}

