% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{pathway_HM}
\alias{pathway_HM}
\title{pathway_HM}
\usage{
pathway_HM(
  mean_NES_df,
  pathway_name,
  pathways,
  sort_by = 1,
  manual_pathway = F
)
}
\arguments{
\item{mean_NES_df}{ipsum...}

\item{pathway_name}{ipsum...}

\item{pathways}{ipsum...}

\item{sort_by}{ipsum...}

\item{manual_pathway}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
