% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_copernicus.r
\name{plot_copernicus}
\alias{plot_copernicus}
\title{This is a function to plot downloaded copernicus products.}
\usage{
plot_copernicus(
  yy,
  mm,
  dd,
  product,
  folder,
  level = 1,
  rangelon = NULL,
  rangelat = NULL,
  rangevar = NULL,
  main = "",
  xlab = "",
  ylab = "",
  log = F,
  interpolate = T
)
}
\arguments{
\item{yy}{year}

\item{mm}{month}

\item{dd}{day}

\item{product}{product}

\item{folder}{folder}

\item{level}{level}

\item{rangelon}{range of longitudes. The default is the whole area in the NetCDF file}

\item{rangelat}{range of latitudes. The default is the whole area in the NetCDF file}

\item{main}{Plot title}

\item{xlab}{x-axis title}

\item{ylab}{y-axis title}

\item{log}{Logical indicating whether to log transform (chl)}

\item{interpolate}{Logical indicating whether to interpolate or show raw data}
}
\description{
This is a function to plot downloaded copernicus products.
}
\examples{
See temperature at around 15m deep.
dir.create('./test')
download_copernicus(yearstart=2014,monthstart=1,yearend=2014,
monthend=1,minlon=-15,maxlon=5,minlat=35,maxlat=45,mindepth=10,maxdepth=20,c('mtemp'),'./test',username='username',
password='mypassword')
test<-load_copernicus(2014,1,'mtemp','./test')
test$depth
plot_copernicus(yy=2014,mm=1,dd=12,product='mtemp',folder='./test',level=3,main='Prueba temperatura a 15 m 12/01/2014',xlab='Longitud',ylab='Latitud',log=T,interpolate = T)
}
\keyword{plot,copernicus}
