% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reaction_network.R
\name{plot_reaction_network}
\alias{plot_reaction_network}
\title{plot_reaction_network}
\usage{
plot_reaction_network(
  network_and_attributes,
  t_table,
  scores_df,
  column_index,
  rbrewer_plalette_name = "RdBu",
  vis.height = 700,
  vis.degree = 2
)
}
\arguments{
\item{network_and_attributes}{lorem ipsum...}

\item{t_table}{lorem ipsum...}

\item{scores_df}{lorem ipsum...}

\item{column_index}{lorem ipsum...}

\item{rbrewer_plalette_name}{lorem ipsum...}
}
\value{
lorem ipsum...
}
\description{
plot_reaction_network
}
