% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_metabolite_enrichment.R
\name{plot_significant_pathways}
\alias{plot_significant_pathways}
\title{\code{plot_significant_pathways}}
\usage{
plot_significant_pathways(enrichmentDF, score)
}
\arguments{
\item{enrichmentDF}{Data frame containing all significant pathways}
}
\value{
Bar plot of the most significant pathways by enrichment score
}
\description{
Function to plot the interesting pathways after metabolite enrichment analysis
}
