% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plotMetaboliteContribution}
\alias{plotMetaboliteContribution}
\title{plotMetaboliteContribution}
\usage{
plotMetaboliteContribution(
  enzyme,
  stat_df,
  metabolite_sets,
  contrast_index,
  stat_name = "stat",
  scaling_factor = 1,
  nLabels = 10
)
}
\arguments{
\item{enzyme}{ipsum...}

\item{stat_df}{ipsum...}

\item{metabolite_sets}{ipsum...}

\item{contrast_index}{ipsum...}

\item{stat_name}{ipsum...}

\item{scaling_factor}{ipsum...}

\item{nLabels}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
