% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metactivity_core_functions.R
\name{prepare_metabolite_set}
\alias{prepare_metabolite_set}
\title{prepare_metabolite_set}
\usage{
prepare_metabolite_set(penalty_range, reaction_tree, metab_list)
}
\arguments{
\item{penalty_range}{ipsum...}

\item{reaction_tree}{ipsum...}

\item{metab_list}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
