% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_metabolite_enrichment.R
\name{rearrange_dataframe}
\alias{rearrange_dataframe}
\title{\code{rearrange_dataframe}}
\usage{
rearrange_dataframe(network)
}
\arguments{
\item{network}{source/target network containing enzymes and metabolites}
}
\value{
data frame with metabolites and their target/source enzymes
}
\description{
Function to separate the complexes into a vector of enzyme strings, map the
identifiers to their enzyme name using recon2_redhuman, and give back the full 
source/target network with enzymes, metabolites and translated complexes.
}
