% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon2_redhuman.R
\docType{data}
\name{recon2_redhuman}
\alias{recon2_redhuman}
\title{recon2_redhuman}
\format{
An object of class \code{list} of length 8.
}
\usage{
data(recon2_redhuman)
}
\description{
recon2_redhuman
}
\examples{
data(recon2_redhuman)

}
\references{
{
  https://pubmed.ncbi.nlm.nih.gov/36539415/
}
}
\keyword{recon2_redhuman}
