% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_pathway.R
\name{remove_cofactors}
\alias{remove_cofactors}
\title{\code{remove_cofactors}}
\usage{
remove_cofactors(reaction_network, compound_list = kegg_compounds)
}
\arguments{
\item{reaction_network}{ipsum...}

\item{compound_list}{KEGGREST coumpound list}
}
\value{
ipsum...
}
\description{
This function removes cofactors from reaction networks generate by the model_to_pathway_sif function
}
