% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{t_table_metactivity_input_formater}
\alias{t_table_metactivity_input_formater}
\title{t_table_metactivity_input_formater}
\usage{
t_table_metactivity_input_formater(
  metabolomic_t_table,
  mapping_table,
  affixes = c("c", "l", "x", "m", "e", "n", "r")
)
}
\arguments{
\item{metabolomic_t_table}{ipsum...}

\item{mapping_table}{ipsum...}

\item{affixes}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
