% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_complexes.R
\name{translate_complexes}
\alias{translate_complexes}
\title{\code{translate_complexes}}
\usage{
translate_complexes(Network)
}
\arguments{
\item{Network}{source/target network containing enzymes, complexes, metabolites}
}
\value{
source/target network containing enzymes, translated complexes, metabolites
}
\description{
Function to separate the complexes into a vector of enzyme strings, map the
identifiers to their enzyme name using recon2_redhuman, and give back the full 
source/target network with enzymes, metabolites and translated complexes.
}
