% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{translate_results}
\alias{translate_results}
\title{translate_results}
\usage{
translate_results(
  regulons_df,
  t_table,
  mapping_table,
  compress_compartments = F
)
}
\arguments{
\item{regulons_df}{ipsum...}

\item{t_table}{ipsum...}

\item{mapping_table}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
