% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{ttop_list_to_t_table}
\alias{ttop_list_to_t_table}
\title{ttop_list_to_t_table}
\usage{
ttop_list_to_t_table(ttop_list)
}
\arguments{
\item{ttop_list}{list of limma top table results obtained from limma_res_to_ttop_list}
}
\value{
data.frame of t-values, where columns are the contrasts and rows are the omic features tested
}
\description{
This function converts a list of limma top table results into a single 
dataframe of t-values. The list of top table results can be obtained using 
the limma_res_to_ttop_list function
}
