% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{ttopFormatter}
\alias{ttopFormatter}
\title{\code{ttopFormatter}}
\usage{
ttopFormatter(ttop)
}
\arguments{
\item{ttop}{a toptable dataframe generated by the topTable function of limma}
}
\value{
a dataframe similar to the output of topTable function of limma but with first column as IDs instead of onyl row.names.
}
\description{
This function is simply designed to format the toptable of limma with first column as gene identifiers instead of only row.names.
}
