% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.copernicusmarine.r
\name{update.copernicusmarine}
\alias{update.copernicusmarine}
\title{Ésta es una función para actualizar el módulo de copernicusmarine}
\usage{
\method{update}{copernicusmarine}()
}
\description{
Ésta es una función para actualizar el módulo de copernicusmarine
}
\examples{
update.copernicusmarine()


}
