\name{validFVCOMGrid}
\alias{validFVCOMGrid}
\title{Check if a \code{fvcom.grid} instance is valid.}
\usage{
  validFVCOMGrid(object)
}
\arguments{
  \item{object}{A \code{fvcom.grid} instance to check for
  validity.}
}
\value{
  \code{TRUE} if object is well formed, otherwise
  \code{FALSE.}
}
\description{
  Check if a \code{fvcom.grid} instance is valid.
}

