% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metactivity_core_functions.R
\name{condense_metabolite_set}
\alias{condense_metabolite_set}
\title{condense_metabolite_set}
\usage{
condense_metabolite_set(
  reaction_set_list,
  condense_sign = T,
  condense_compartments = F,
  ignore_sign = F
)
}
\arguments{
\item{reaction_set_list}{list, metabolic-enzyme set collections as returned by the prepare_metabolite_set function}

\item{condense_sign}{boolean, summarise upstream and downstream position of metabolites into a signle relative position}

\item{condense_compartments}{boolean, wether the compartment information of hte metabolic network sohuld be ignored}

\item{ignore_sign}{boolean, wether the upstream/downstream position of metabolic should be ignored}
}
\value{
a metabolic-enzyme set collections
}
\description{
this function condense the metabolic-enzyme set collections to summarise ignore specific types of
information such as compartments of relative positions (sign)
}
