% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{limma_res_to_ttop_list}
\alias{limma_res_to_ttop_list}
\title{limma_res_to_ttop_list}
\usage{
limma_res_to_ttop_list(limma_res, comp_names, number, adjust.method = "fdr")
}
\arguments{
\item{limma_res}{ipsum...}

\item{comp_names}{ipsum...}

\item{number}{ipsum...}

\item{adjust.method}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
