% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{makeContrastsAlt}
\alias{makeContrastsAlt}
\title{makeContrastsAlt}
\usage{
makeContrastsAlt(targets, comparisons)
}
\arguments{
\item{targets}{ipsum...}

\item{comparisons}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
