% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_pathway.R
\name{model_to_pathway_sif}
\alias{model_to_pathway_sif}
\title{\code{model_to_pathway_sif}}
\usage{
model_to_pathway_sif(
  pathway_to_keep,
  pathways = recon2_redhuman$pathway,
  rxns = recon2_redhuman$rxns,
  rules = recon2_redhuman$rules,
  minmax = recon2_redhuman$minmax,
  stochio = recon2_redhuman$stochio,
  metab = recon2_redhuman$metab,
  mapping = recon2_redhuman$gene_mapping,
  recon1_metabolites = recon2_redhuman$metab_mapping
)
}
\arguments{
\item{pathway_to_keep}{character vector with the ID of the pathways that 
should be used to build the SIF reaction network. 
Use "View(unique(recon2_redhuman$pathway))" to see available pathway IDs}

\item{pathways}{pathway corresponding to redhuman reaction indexes}

\item{rxns}{reaction IDs corresponding to redhuman reaction indexes}

\item{rules}{gene rules corresponding to redhuman reaction indexes}

\item{minmax}{min max flux boundaries corresponding to redhuman reaction indexes}

\item{stochio}{redhuman stochiometric matrix}

\item{metab}{metabolite IDs corresponding to redhuman metabolite indexes}

\item{mapping}{obsolete}

\item{recon1_metabolites}{obsolete}
}
\value{
a subreaction network in sif format
}
\description{
This function generate a subreaction network in sif format from the 
recon_redhuman model for a given set of pathways
}
