% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metactivity_core_functions.R
\name{prepare_metabolite_set}
\alias{prepare_metabolite_set}
\title{prepare_metabolite_set}
\usage{
prepare_metabolite_set(penalty_range, forest, measured_metabolites)
}
\arguments{
\item{penalty_range}{numeric vector, between minimum 1 and maximum 9}

\item{forest}{list, a forest representation of a metabolic network.}

\item{measured_metabolites}{character vector, names of measured metabolites}
}
\value{
a series of metabolic-enzyme set collections
}
\description{
This functions create a series of metabolic-enzyme set collections over a range of penalties
(see tutorial)
}
