% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metactivity_core_functions.R
\name{prepare_regulon_df}
\alias{prepare_regulon_df}
\title{condense_metabolite_set}
\usage{
prepare_regulon_df(reaction_set_list_merged, penalty)
}
\arguments{
\item{reaction_set_list_merged}{ipsum...}

\item{penalty}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
