% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metactivity_core_functions.R
\name{prepare_regulon_df}
\alias{prepare_regulon_df}
\title{prepare_regulon_df}
\usage{
prepare_regulon_df(reaction_set_list_merged, penalty)
}
\arguments{
\item{reaction_set_list_merged}{a metabolic-enzyme set collections, 
as returned by the condense_metabolite_set function}

\item{penalty}{numeric, the index of the reaction_set_list_merged corresponding to
the desired penalty}
}
\value{
a regulon dataframe with three columns correposning to the enzme, the target metabolite and the weight
}
\description{
extract and cleanup the regulon collection correpsonding to the desired penalty
}
