% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_metabolite_enrichment.R
\name{rearrange_dataframe}
\alias{rearrange_dataframe}
\title{\code{rearrange_dataframe}}
\usage{
rearrange_dataframe(network)
}
\arguments{
\item{network}{source/target network containing enzymes and metabolites}
}
\value{
data frame with metabolites and their source/target enzymes
}
\description{
Function to rearrange the data frame to obtain an enzyme/metabolite df
instead of a source/target df, i.e. all enzymes are in one column
and all metabolites in the second column.
}
