% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{runLimma}
\alias{runLimma}
\title{runLimma}
\usage{
runLimma(measurements, targets, comparisons = NULL, regress_out = NULL)
}
\arguments{
\item{targets}{ipsum...}

\item{comparisons}{ipsum...}

\item{regress_out}{ipsum...}

\item{measurments}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
