% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{t_table_metactivity_input_formater}
\alias{t_table_metactivity_input_formater}
\title{t_table_metactivity_input_formater}
\usage{
t_table_metactivity_input_formater(
  metabolomic_t_table,
  mapping_table,
  affixes = c("c", "l", "x", "m", "e", "n", "r")
)
}
\arguments{
\item{metabolomic_t_table}{the metabolomic t-table (or log2FC table) that 
is returned by ttop_list_to_t_table (or ttop_list_to_log2FC_table)}

\item{mapping_table}{a mapping table to translate the metabolite 
identifiers of the t-table into KEGG IDs}

\item{affixes}{a character vectors corresponding to the compartment affixes
to be considered. If you wish to only focus on one compartment, 
e.g. mitochondria, you can input only c("m") for example.}
}
\value{
a formatted t-table (or log2FC table) for use with the metactivity function
}
\description{
This function allows to format a table of t-values (or log2FCs) for the 
metactivity function. Essentially, it adds a compartment code (using BIGG
nomenclature) at the end of each metabolite identifier, so that they can be 
mapped on the compartmentalized reaction prior knowledge network.
}
