% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metactivity_core_functions.R
\name{target_set_from_forest_2}
\alias{target_set_from_forest_2}
\title{target_set_from_forest_2}
\usage{
target_set_from_forest_2(forest, measured_species, penalty = 0.5, verbose = T)
}
\arguments{
\item{forest}{ipsum...}

\item{measured_species}{ipsum...}

\item{penalty}{ipsum...}

\item{verbose}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
