% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{ttop_list_to_log2FC_table}
\alias{ttop_list_to_log2FC_table}
\title{ttop_list_to_log2FC_table}
\usage{
ttop_list_to_log2FC_table(ttop_list)
}
\arguments{
\item{ttop_list}{list of limma top table results obtained from limma_res_to_ttop_list}
}
\value{
data.frame of t-values, where columns are the contrasts and rows are the omic features tested
}
\description{
same as ttop_list_to_t_table, but with log2FC instead of t-values
}
