% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_aux_functions.R
\name{ttop_list_to_t_table}
\alias{ttop_list_to_t_table}
\title{ttop_list_to_t_table}
\usage{
ttop_list_to_t_table(ttop_list)
}
\arguments{
\item{ttop_list}{ipsum...}
}
\value{
ipsum...
}
\description{
ipsum...
}
